refactor: . Move globalGetFileCache to CacheReader.ts
